bug fixing of confirmation screen and input of large numbers

+++ frontend/app/src/main/java/com/lamba/app/data/statistics/StatisticsModels.kt
@@ -34,7 +34,7 @@ data class StatisticsMetricResponse(
 
 data class StatisticsChartPointResponse(
     val label: String,
-    val value: Int
+    val value: Long
 )
 
 data class StatisticsCategoryResponse(
